src/requirements_news.txt
requests~=2.31.0
scikit-learn
scipy
telegraph
pandas
datetime
beautifulsoup4
python-dotenv